Revert "Merge tizen_build_info.dart back into tizen_builder.dart"
This reverts commit 2814cf8f5ddf404f565d107e6daf5f1d2c8f80bb.
