+++ lib/tizen_builder.dart
@@ -31,6 +31,7 @@ import 'package:process/process.dart';
 
 import 'build_targets/application.dart';
 import 'build_targets/package.dart';
+import 'tizen_build_info.dart';
 import 'tizen_project.dart';
 import 'tizen_sdk.dart';
 import 'tizen_tpk.dart';
@@ -39,21 +40,6 @@ import 'tizen_tpk.dart';
 const String kDeviceProfile = 'DeviceProfile';
 
 TizenBuilder? get tizenBuilder => context.get<TizenBuilder>();
-
-/// See: [AndroidBuildInfo] in `build_info.dart`
-class TizenBuildInfo {
-  const TizenBuildInfo(
-    this.buildInfo, {
-    required this.targetArch,
-    required this.deviceProfile,
-    this.securityProfile,
-  });
-
-  final BuildInfo buildInfo;
-  final String targetArch;
-  final String deviceProfile;
-  final String? securityProfile;
-}
 
 /// See:
 /// - [AndroidGradleBuilder.buildApk] in `gradle.dart`
