+++ lib/tizen_device.dart
@@ -28,6 +28,7 @@ import 'package:path/path.dart';
 import 'package:process/process.dart';
 
 import 'forwarding_log_reader.dart';
+import 'tizen_build_info.dart';
 import 'tizen_builder.dart';
 import 'tizen_sdk.dart';
 import 'tizen_tpk.dart';
